Comprehensive email condition handling + enhanced security
- Added a bool that catches whether 'WA DISP' submitted this report AND
whether if 'WA DISP' has submitted it before. If it not submitted
before, the PDF is emailed to the client.
- Added a new element to clearance evaluation and changing $status
flag. Status will not change if the old status was at OR below the
user's clearance level. This ensure that the proper status is given
once, and not changed by linear progression.